feat: add documentation social cards
