+++ docs/src/app/_social-card/assets/nunito-bold.ttf
